update(AllAlbums):make-styles-shared
