+++ web/apps/photos/src/components/CollectionSelector.tsx
@@ -1,25 +1,37 @@
+import {
+    CreateAlbumTile,
+    CollectionDialogSearchField as SearchField,
+    CollectionTileButton as TileButton,
+    CollectionTileTextOverlay as TopGradientOverlay,
+} from "@/components/CollectionDialog/Primitives";
+import {
+    collectionDialogFullScreenQuery,
+    collectionDialogBodyMutedSx as countSx,
+    collectionDialogDividerSx as dividerSx,
+    collectionDialogHeaderActionsSx as headerActionsSx,
+    collectionDialogHeaderRowSx as headerRowSx,
+    collectionDialogHeaderSx as headerSx,
+    collectionDialogIconButtonSx as iconButtonSx,
+    collectionDialogNoResultsSx as noResultsSx,
+    collectionDialogPaperSx as paperSx,
+    collectionDialogTitleSx as titleSx,
+} from "@/components/CollectionDialog/styles";
 import { CollectionsSortOptions } from "@/components/CollectionsSortOptions";
-import AddIcon from "@mui/icons-material/Add";
 import CloseIcon from "@mui/icons-material/Close";
 import FavoriteIcon from "@mui/icons-material/Favorite";
 import PushPinIcon from "@mui/icons-material/PushPin";
-import SearchIcon from "@mui/icons-material/Search";
 import {
     Box,
     Dialog,
     IconButton,
-    InputBase,
     Stack,
-    styled,
     Tooltip,
     Typography,
     useMediaQuery,
-    type SxProps,
-    type Theme,
 } from "@mui/material";
 import type { ModalVisibilityProps } from "ente-base/components/utils/modal";
 import type { Collection } from "ente-media/collection";
-import { BaseTileButton, ItemCard } from "ente-new/photos/components/Tiles";
+import { ItemCard } from "ente-new/photos/components/Tiles";
 import {
     canAddToCollection,
     canMoveToCollection,
@@ -72,7 +84,7 @@ export const CollectionSelector: React.FC<CollectionSelectorProps> = (
         collectionForCollectionSummaryID,
     } = props;
 
-    const isFullScreen = useMediaQuery("(max-width: 490px)");
+    const isFullScreen = useMediaQuery(collectionDialogFullScreenQuery);
 
     const [searchTerm, setSearchTerm] = useState("");
     const [sortBy, setSortBy] =
@@ -281,73 +293,6 @@ const useCollectionSelectorSortByLocalState = (
     return [value, setter] as const;
 };
 
-const surfaceStroke = "#e0e0e0";
-const surfaceStrokeDark = "rgba(255 255 255 / 0.12)";
-
-const paperSx: SxProps<Theme> = (theme) => ({
-    width: "min(500px, calc(100svw - 32px))",
-    maxWidth: "500px",
-    boxSizing: "content-box",
-    borderRadius: "20px",
-    border: `1px solid ${surfaceStroke}`,
-    backgroundColor: "#f4f4f4",
-    backgroundImage: "none",
-    boxShadow: "none",
-    color: "text.base",
-    "@media (min-width: 491px)": { height: "100%" },
-    "@media (max-width: 490px)": {
-        width: "100%",
-        maxWidth: "100%",
-        height: "100%",
-        boxSizing: "border-box",
-        borderRadius: 0,
-        border: "none",
-    },
-    ...theme.applyStyles("dark", {
-        borderColor: surfaceStrokeDark,
-        backgroundColor: "#1b1b1b",
-    }),
-});
-
-const headerSx = { p: "20px", gap: "16px" };
-const headerRowSx = {
-    alignItems: "flex-start",
-    justifyContent: "space-between",
-    gap: "12px",
-};
-const headerActionsSx = { alignItems: "center", gap: 1, flexShrink: 0 };
-const titleSx = {
-    fontSize: 24,
-    lineHeight: "32px",
-    fontWeight: 600,
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-    whiteSpace: "nowrap",
-};
-const countSx = {
-    fontSize: 14,
-    lineHeight: "20px",
-    fontWeight: 500,
-    color: "text.muted",
-};
-const iconButtonSx = (theme: Theme) => ({
-    width: 38,
-    height: 38,
-    p: 0,
-    color: "text.base",
-    backgroundColor: "background.paper",
-    "&:hover": { backgroundColor: "fill.faintHover" },
-    ...theme.applyStyles("dark", {
-        backgroundColor: "rgba(255 255 255 / 0.12)",
-    }),
-});
-const dividerSx = (theme: Theme) => ({
-    height: "1px",
-    backgroundColor: "rgba(0 0 0 / 0.06)",
-    ...theme.applyStyles("dark", {
-        backgroundColor: "rgba(255 255 255 / 0.08)",
-    }),
-});
 const gridSx = {
     flex: 1,
     minHeight: 0,
@@ -359,35 +304,6 @@ const gridSx = {
     alignContent: "start",
     overflowY: "auto",
 };
-const noResultsSx = {
-    flex: 1,
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-    minHeight: 154,
-};
-
-const TileButton = styled(BaseTileButton)`
-    width: 100%;
-    aspect-ratio: 1;
-    height: auto;
-    border-radius: 16px;
-`;
-
-const TopGradientOverlay = styled("div")`
-    position: absolute;
-    inset: 0;
-    padding: 10px;
-    color: #fff;
-    background: linear-gradient(
-        -10deg,
-        rgba(0, 0, 0, 0.1) 0%,
-        rgba(0, 0, 0, 0.2) 50%,
-        rgba(0, 0, 0, 0.4) 60%,
-        rgba(0, 0, 0, 0.6) 100%
-    );
-`;
-
 interface CollectionSummaryButtonProps {
     collectionSummary: CollectionSummary;
     onClick: (collectionSummaryID: number) => void;
@@ -447,86 +363,3 @@ const CollectionSummaryButton: React.FC<CollectionSummaryButtonProps> = ({
         </ItemCard>
     );
 };
-
-const CreateAlbumTile: React.FC<{ onClick: () => void }> = ({ onClick }) => (
-    <TileButton
-        aria-label={t("new_album")}
-        onClick={onClick}
-        sx={{ borderRadius: "16px" }}
-    >
-        <CreateTileInner>
-            <AddIcon />
-        </CreateTileInner>
-    </TileButton>
-);
-
-const CreateTileInner = styled("span")(({ theme }) => ({
-    position: "absolute",
-    inset: 0,
-    padding: 10,
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    border: "1px dashed",
-    borderColor: theme.vars.palette.stroke.muted,
-    borderRadius: 16,
-    color: theme.vars.palette.text.muted,
-    "&:hover": { borderColor: "rgba(0 0 0 / 0.45)" },
-    ...theme.applyStyles("dark", {
-        "&:hover": { borderColor: "rgba(255 255 255 / 0.45)" },
-    }),
-}));
-
-interface SearchFieldProps {
-    value: string;
-    onChange: (value: string) => void;
-}
-
-const SearchField: React.FC<SearchFieldProps> = ({ value, onChange }) => (
-    <Stack direction="row" sx={searchFieldSx}>
-        <SearchIcon sx={searchIconSx} />
-        <InputBase
-            fullWidth
-            autoFocus
-            placeholder={t("albums_search_hint")}
-            value={value}
-            onChange={(e) => onChange(e.target.value)}
-            sx={{
-                fontSize: 14,
-                lineHeight: "20px",
-                fontWeight: 500,
-                color: "text.base",
-                "& input::placeholder": { color: "text.muted", opacity: 1 },
-            }}
-            endAdornment={
-                value ? (
-                    <CloseIcon
-                        fontSize="small"
-                        onClick={() => onChange("")}
-                        sx={{
-                            color: "stroke.muted",
-                            cursor: "pointer",
-                            "&:hover": { color: "text.base" },
-                        }}
-                    />
-                ) : undefined
-            }
-        />
-    </Stack>
-);
-
-const searchFieldSx = (theme: Theme) => ({
-    alignItems: "center",
-    gap: "10px",
-    height: 44,
-    borderRadius: "16px",
-    backgroundColor: "background.paper",
-    px: "14px",
-    ...theme.applyStyles("dark", { backgroundColor: "#282828" }),
-});
-const searchIconSx = (theme: Theme) => ({
-    fontSize: 20,
-    flexShrink: 0,
-    color: "rgba(0 0 0 / 0.4)",
-    ...theme.applyStyles("dark", { color: "rgba(255 255 255 / 0.4)" }),
-});
